split SelectPipetteModal into 4 fns
