+++ protocol-designer/src/components/organisms/SelectPipetteModal/SelectPipetteTips.tsx
@@ -1,0 +1,161 @@
+import { useTranslation } from 'react-i18next'
+import styled, { css } from 'styled-components'
+import { useDispatch, useSelector } from 'react-redux'
+import {
+  ALIGN_CENTER,
+  ALIGN_STRETCH,
+  Box,
+  Btn,
+  Checkbox,
+  COLORS,
+  CURSOR_POINTER,
+  DIRECTION_COLUMN,
+  DISPLAY_FLEX,
+  DISPLAY_INLINE_BLOCK,
+  Flex,
+  PRODUCT,
+  SPACING,
+  StyledText,
+  TYPOGRAPHY,
+  WRAP,
+} from '@opentrons/components'
+import { FLEX_ROBOT_TYPE } from '@opentrons/shared-data'
+import { removeOpentronsPhrases } from '../../../utils'
+import { getAllowAllTipracks } from '../../../feature-flags/selectors'
+import { createCustomTiprackDef } from '../../../labware-defs/actions'
+import { setFeatureFlags } from '../../../feature-flags/actions'
+import { useKitchen } from '../Kitchen/hooks'
+
+import type { Dispatch, SetStateAction } from 'react'
+import type { UseFormSetValue } from 'react-hook-form'
+import type { PipetteMount, RobotType } from '@opentrons/shared-data'
+import type { ThunkDispatch } from 'redux-thunk'
+import type { BaseState } from '../../../types'
+
+interface SelectPipetteTipsProps {
+  mount: PipetteMount
+  robotType: RobotType
+  tiprackOptions: Record<string, string>
+  setValue: UseFormSetValue<any>
+  selectedValues: string[]
+  pipetteVolume: string | null
+  setIncompatibleTip: Dispatch<SetStateAction<boolean>>
+}
+
+const MAX_TIPRACKS_ALLOWED = 3
+
+export function SelectPipetteTips(props: SelectPipetteTipsProps): JSX.Element {
+  const {
+    mount,
+    robotType,
+    tiprackOptions,
+    setValue,
+    selectedValues,
+    pipetteVolume,
+    setIncompatibleTip,
+  } = props
+  const { t } = useTranslation('create_new_protocol')
+  const { makeSnackbar } = useKitchen()
+  const allowAllTipracks = useSelector(getAllowAllTipracks)
+  const dispatch = useDispatch<ThunkDispatch<BaseState, any, any>>()
+
+  return (
+    <Flex flexDirection={DIRECTION_COLUMN} gridGap={SPACING.spacing4}>
+      <Flex flexDirection={DIRECTION_COLUMN} gridGap={SPACING.spacing16}>
+        <StyledText desktopStyle="headingSmallBold">
+          {t('pipette_tips')}
+        </StyledText>
+        <Box
+          css={css`
+            gap: ${SPACING.spacing4};
+            display: ${DISPLAY_FLEX};
+            flex-wrap: ${WRAP};
+            align-items: ${ALIGN_CENTER};
+            align-content: ${ALIGN_CENTER};
+            align-self: ${ALIGN_STRETCH};
+          `}
+        >
+          {Object.entries(tiprackOptions).map(([value, name]) => (
+            <Checkbox
+              key={value}
+              isChecked={selectedValues.includes(value)}
+              labelText={removeOpentronsPhrases(name)}
+              onClick={() => {
+                const isCurrentlySelected = selectedValues.includes(value)
+
+                if (isCurrentlySelected) {
+                  setValue(
+                    `pipettesByMount.${mount}.tiprackDefURI`,
+                    selectedValues.filter(v => v !== value)
+                  )
+                } else {
+                  if (selectedValues.length === MAX_TIPRACKS_ALLOWED) {
+                    makeSnackbar(t('up_to_3_tipracks') as string)
+                  } else {
+                    setValue(`pipettesByMount.${mount}.tiprackDefURI`, [
+                      ...selectedValues,
+                      value,
+                    ])
+                  }
+                }
+              }}
+            />
+          ))}
+          <StyledLabel>
+            <StyledText
+              desktopStyle="bodyDefaultRegular"
+              padding={SPACING.spacing4}
+            >
+              {t('add_custom_tips')}
+            </StyledText>
+            <input
+              data-testid="SelectPipettes_customTipInput"
+              type="file"
+              onChange={e => dispatch(createCustomTiprackDef(e))}
+            />
+          </StyledLabel>
+          {pipetteVolume === 'p1000' && robotType === FLEX_ROBOT_TYPE ? null : (
+            <Btn
+              onClick={() => {
+                if (allowAllTipracks) {
+                  dispatch(
+                    setFeatureFlags({
+                      OT_PD_ALLOW_ALL_TIPRACKS: !allowAllTipracks,
+                    })
+                  )
+                } else {
+                  setIncompatibleTip(true)
+                }
+              }}
+              textDecoration={TYPOGRAPHY.textDecorationUnderline}
+            >
+              <StyledLabel>
+                <StyledText
+                  desktopStyle="bodyDefaultRegular"
+                  padding={SPACING.spacing4}
+                >
+                  {allowAllTipracks
+                    ? t('show_default_tips')
+                    : t('show_all_tips')}
+                </StyledText>
+              </StyledLabel>
+            </Btn>
+          )}
+        </Box>
+      </Flex>
+    </Flex>
+  )
+}
+
+const StyledLabel = styled.label`
+  text-decoration: ${TYPOGRAPHY.textDecorationUnderline};
+  font-size: ${PRODUCT.TYPOGRAPHY.fontSizeBodyDefaultSemiBold};
+  display: ${DISPLAY_INLINE_BLOCK};
+  cursor: ${CURSOR_POINTER};
+  input[type='file'] {
+    display: none;
+  }
+  &:hover {
+    color: ${COLORS.blue50};
+  }
+`
