+++ protocol-designer/src/components/organisms/SelectPipetteModal/SelectPipetteType.tsx
@@ -1,0 +1,70 @@
+import { useTranslation } from 'react-i18next'
+import {
+  DIRECTION_COLUMN,
+  Flex,
+  RadioButton,
+  SPACING,
+  StyledText,
+  WRAP,
+} from '@opentrons/components'
+import { PIPETTE_TYPES } from '../../../pages/Onboarding/constants'
+import type { Dispatch, SetStateAction } from 'react'
+import type { UseFormSetValue } from 'react-hook-form'
+import type { Gen, PipetteType } from '../../../pages/Onboarding/types'
+import type { PipetteMount, RobotType } from '@opentrons/shared-data'
+import type { FormPipettesByMount } from '../../../step-forms'
+
+interface SelectPipetteTypeProps {
+  mount: PipetteMount
+  robotType: RobotType
+  pipettesByMount: FormPipettesByMount
+  setPipetteGen: Dispatch<SetStateAction<'flex' | Gen>>
+  setPipetteVolume: Dispatch<SetStateAction<string | null>>
+  setPipetteType: Dispatch<SetStateAction<PipetteType | null>>
+  pipetteType: PipetteType | null
+  setValue: UseFormSetValue<any>
+}
+
+export function SelectPipetteType(props: SelectPipetteTypeProps): JSX.Element {
+  const {
+    mount,
+    robotType,
+    pipettesByMount,
+    setPipetteGen,
+    setPipetteVolume,
+    setPipetteType,
+    pipetteType,
+    setValue,
+  } = props
+  const { t } = useTranslation('create_new_protocol')
+
+  return (
+    <Flex flexDirection={DIRECTION_COLUMN} gridGap={SPACING.spacing12}>
+      <StyledText desktopStyle="headingSmallBold">
+        {t('pipette_type')}
+      </StyledText>
+      <Flex gridGap={SPACING.spacing4} flexWrap={WRAP}>
+        {PIPETTE_TYPES[robotType].map(type => {
+          return type.value === '96' &&
+            (mount === 'right' ||
+              (mount === 'left' &&
+                pipettesByMount.right.pipetteName != null)) ? null : (
+            <RadioButton
+              key={`${type.label}_${type.value}`}
+              onChange={() => {
+                setPipetteType(type.value)
+                setPipetteGen('flex')
+                setPipetteVolume(null)
+                setValue(`pipettesByMount.${mount}.pipetteName`, undefined)
+                setValue(`pipettesByMount.${mount}.tiprackDefURI`, undefined)
+              }}
+              buttonLabel={t(`shared:${type.label}`)}
+              buttonValue="single"
+              isSelected={pipetteType === type.value}
+            />
+          )
+        })}
+      </Flex>
+    </Flex>
+  )
+}
